Implement APIClient class
